pyproject.toml
[tool.poetry]
name = "api-project-generator"
version = "0.5.2"
description = ""
authors = ["Gustavo Correa <self.gustavocorrea@gmail.com>"]
repository = "https://github.com/guscardvs/api-project-generator"
readme = "README.md"

[tool.poetry.scripts]
api-project = "api_project_generator.main:app"

[tool.poetry.dependencies]
python = "^3.9"
GitPython = "^3.1.18"
gyver = "^2.3.3"
sqlalchemy = "^2.0.17"
tomlkit = "^0.11.8"
jinja2 = "^3.1.2"
typer = "^0.9.0"
appdirs = "^1.4.4"

[tool.poetry.group.style.dependencies]
black = "^23.7.0"
autoflake = "^2.2.0"
isort = "^5.12.0"
pylint = "^2.17.4"

[tool.isort]
profile = "black"

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.5.2"
tag_format = "$version"
version_files = ["api_project_generator/__init__.py"]
update_changelog_on_bump = true


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
